+++ CHANGELOG.md
@@ -7,10 +7,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## Unreleased
 [Compare v3.4.0 - Unreleased](https://github.com/exonet/powerdns-php/compare/v3.4.0...develop)
 
-## [v3.4.0](https://github.com/exonet/powerdns-php/releases/tag/v3.4.0) - 2021-10-06
+## [v3.4.0](https://github.com/exonet/powerdns-php/releases/tag/v3.4.0) - 2021-10-19
 [Compare v3.3.1 - v3.4.0](https://github.com/exonet/powerdns-php/compare/v3.3.1...v3.4.0)
 ### Added
-- The `overwriteConnector` method in the Powerdns class.
+- The `setConnector` method in the Powerdns class.
 
 ### Changed
 - The private methods in the `Connector` class are now protected instead of private.